Add a timeout to notification icon image loading.
This CL adds a new timeout to the NotificationImageLoader which will
trigger a timeout instead of hanging a service worker until the base
service worker timeout happens. This will allow the code to recover from
loading errors and better avoid the dreaded default notifications. The
timeout for images will be 90 seconds, which currently covers 99.9% of
all image loads.

Testing was also added to verify the timeout trigger.

BUG=579137

Review-Url: https://chromiumcodereview.appspot.com/2432783003
Cr-Commit-Position: refs/heads/master@{#426768}