mash: Introduce ShowWebDialogWithContainer for webui dialogs
Several webui dialogs spawned by the ash system tray can be shown either as
children of the general settings webui window or by themselves at the login
screen. In the latter case we need to place them in the right ash window
container. Under mash, chrome is not aware of the ash window hierarchy, so
it needs to use a container ID rather than a native window parent.

Wire up the "set time" dialog as an example.

BUG=657021
TEST=manual, spawn set time dialog from settings and from lock screen
TBR=stevenjb@chromium.org for function rename touching c/b/ui/webui/chromeos

Review-Url: https://chromiumcodereview.appspot.com/2441133002
Cr-Commit-Position: refs/heads/master@{#426953}