Record UMA metrics for Must-Staple certificates on private roots
We'd like to get an idea of what MITM proxies tend to do with the TLS
Feature Extension (colloquially known as Must-Staple). If MITM proxies
blindly copy the extension into generated certificates, then deploying
Must-Staple will cause a lot of breakage, due to MITM proxies generating
Must-Staple certificates but not stapling OCSP responses.

This CL adds an UMA metric for the presence of the TLS Feature Extension
in certificates that chain to private roots, as a baby step in this
investigation. (Note that this conflates misbehaving MITM proxies with
private PKIs that are using Must-Staple, so it's only a starting point
for an investigation.)

A new asn1::HasTLSFeatureExtension() function is used to record this
histogram.

BUG=633732

Review-Url: https://chromiumcodereview.appspot.com/2436233002
Cr-Commit-Position: refs/heads/master@{#426971}