Change logging of specific tests and builds to rebaseline to debug level.
This is a non-urgent minor change in `webkit-patch rebaseline-cl` that would make it a little less verbose. Right now when you run `webkit-patch rebaseline-cl` by default it prints the list of tests to rebaseline twice: once before starting, listing specific builds, and once when actually downloading baselines.

After this CL, the more specific information (which builds are used) is only printed when --verbose is passed, and the default behavior is less verbose.

Review-Url: https://codereview.chromium.org/2438393002
Cr-Commit-Position: refs/heads/master@{#427004}