AssociatedURLLoader shouldn't derive from WebURLLoader
The AssociatedURLLoader is derived from the WebURLLoader interface while
WebURLLoaderImpl is also derived from it. WebURLLoaderImpl is an API to
use the resource dispatcher and net stack beyond it. AssociatedURLLoader
is a wrapper for the DocumentThreadableLoader logic which implements
CORS, etc. defined in the Fetch Standard. They're not as-is. Even
looking at the API, there're clear differences e.g. the
WebURLLoaderOptions is specific to the AssociatedURLLoader.

This unnecessary inheritance has been causing confusion when reading
the code. We should remove this inheritance.

R=bbudge@chromium.org,jochen@chromium.org,kinuko@chromium.org,raymes@chromium.org,dalecurtis@chromium.org,mmenke@chromium.org
BUG=none

Review-Url: https://codereview.chromium.org/2399463007
Cr-Commit-Position: refs/heads/master@{#427022}