Fix possible multiple calls to MidiManager::StartInitialization() in StartSession()
We assume the caller is the first client and call StartInitialization() if
|pending_clients_| is empty, however |pending_clients_| may be cleared before
initialization finishes and cause successive StartSession() calls to call
StartInitialization() again. This leads to unwanted behavior depending on
platform-specific implementation.

Now we use an enum field to track initialization state, and flatten logic in
StartSession() to make things simple.

BUG=658662

Review-Url: https://codereview.chromium.org/2443113002
Cr-Commit-Position: refs/heads/master@{#427047}