Make "to allow pairing" not be a link in the WebBluetooth chooser
When the Bluetooth adapter is off, the chooser will show:
"Turn on Bluetooth to allow pairing", the whole text is a link.
This CL modifies code to make only "Turn on Bluetooth" a link.
And the "to allow pairing" text is not a link.

I uploaded some screenshots in the issue page.

BUG=658094

Review-Url: https://codereview.chromium.org/2439453006
Cr-Commit-Position: refs/heads/master@{#427116}