Make blink setting/feature names case-sensitive
Previously we match blink setting/feature names ignoring case. This
causes trouble when the name is used in a case-sensitive environment,
e.g. Linux file system. Make them case-sensitive to avoid the trouble.

third_party/WebKit/LayoutTests/FlagExpectations/enable-blink-features=slimmingPaintInvalidation
which is a violation of the new case-sensitive rule is removed in this
CL. It will be re-added in https://codereview.chromium.org/2448633002/
to avoid problems on platforms using case-insensitive file names.

Review-Url: https://codereview.chromium.org/2445123002
Cr-Commit-Position: refs/heads/master@{#427219}