Show/Hide the shelf when showing/hiding IME menu.
If the shelf is auto-hidden, show the opt-in IME menu with accelerator 'Alt+Shilf+K', the IME menu bubble cannot be oped right (it opens half off the screen, shifted to the right).
We should notify to bring out the shelf before showing the IME menu.

When closing the opt-in IME menu with 'Esc' key, the shelf cannot be auto hidden immediately. We should notify the auto hide state changed to auto hide the shelf.

BUG=642423
TEST=Verified on local build.

Review-Url: https://codereview.chromium.org/2432163006
Cr-Commit-Position: refs/heads/master@{#427223}