[wrapper-tracing] Fix write barrier within base constructors
We potentially call |ScriptWrappable::writeBarrier(object, value)| with a
|value| that is still in construction, making it impossible to call virtual
methods. For GarbageCollectedMixin however, we would require doing so. Mitigate
this scenario by just pushing all infos on the marking deque. Assuming that
AdvanceTracing is not called in such scenarios, the next time we drain the
deque, all required callbacks are properly set.

BUG=chromium:468240
R=haraken@chromium.org,hlopko@chromium.org

Review-Url: https://codereview.chromium.org/2440323003
Cr-Commit-Position: refs/heads/master@{#427355}