Android accessibility: fix word/character navigation in unfocused text fields
We have two different code paths for word / character granularity movement:
1. For static text, we update our internal state to keep track of a
   fake cursor
2. For editable text, we update the actual selection

The purpose of #2 was to allow word/character navigation while editing
a text field, but we weren't properly handling the case where it's an
editable text field that isn't currently focused. The fix is to make sure
we use code path #1 when navigating in a text field that isn't currently
focused.

BUG=656782

Review-Url: https://codereview.chromium.org/2420413006
Cr-Commit-Position: refs/heads/master@{#427392}