Create test harness for comparing Syncable strategies.
This CL defines a test harness that we can use for benchmarking
and analysis of different Syncable dsolution types. Scenarios are
coded in an implementation-agnostic manner so that they can be
executed side by side on different solution types, which makes
apples-to-apples performance comparisons a cinch.

We will analyze the qualitative aspects of the coded solution
by focusing on the SyncableConfig subclasses and their dependencies.

R=steimel@chromium.org,lethalantidote@chromium.org
CC=wez@chromium.org

Review-Url: https://codereview.chromium.org/2445523002
Cr-Commit-Position: refs/heads/master@{#427421}