[NTP] Merge the context menu code for NTP tiles and suggestions
Adds a ContextMenuHandler class that takes care of creating the
menu and triages the item clicks, allowing to remove that from
the NTPManager and the Snippet ViewHolder.

Also adds the possibility to disable touch on the legacy ScrollView,
so that we can disable scrolling while the context menu is shown.

BUG=651045

Review-Url: https://codereview.chromium.org/2452543005
Cr-Commit-Position: refs/heads/master@{#427669}