[NoStatePrefetch] Kill renderer after preload scanning
The design doc is linked in the first bug. Main goal: remove Prerender.

Currently NoState Prefetch uses the same logic to manage renderers as the
PrerenderManager. Prefetch code is still living in PrerenderManager, separating
the code from there is planned after we finalize on how much of the
functionality of PrerenderManager is needed.

This is the first tweak in the renderer lifetime: the renderer asks to be killed
as soon as the main resource is fully preload-scanned and all possible
subresources are requested, special Prerendering FinalStatus is recorded to allow:
* testing
* provide a hint that a new prefetch can be started soon

Browsertest changes:

* TestPrerender now keeps FinalStatus for longer, to be able to verify that it
  is correct even after the TestPrerenderContents is destroyed.

* Prefetch browsertests wait only for creation of PrerenderContents, Prerender
  tests have more waiting for page loads on top of that

* PrerenderTestUrlImpl is removed to eliminate one hop through protected method
  in the parent class, which also simplified the prefetch side of the tests: no
  need to mention how many page loads to wait for

* Disabled two tests that load a non-HTML document, correctly killing the
  renderer will be done in later changes.

BUG=632361,649632

Review-Url: https://codereview.chromium.org/2411863002
Cr-Commit-Position: refs/heads/master@{#427678}