Updated the SquareInkDropRipple to be able to animate the center point.
This change makes it possible to show the SquareInkDropRipple at an
initial center point, and then animate it to a target center point. This
will help Views using the SquareInkDropRipple to provide initial
feedback based on click/touch event and give a good indication of the
actual targetable area by animating the ripple to the view's center
point.

BUG=649734

Review-Url: https://codereview.chromium.org/2449093002
Cr-Commit-Position: refs/heads/master@{#428053}