Refactoring: Move _port_skips_test to Port class.
As part of rebaselining, when copying existing baselines, we don't want to
make new platform-specific baseline files for platforms where tests are
skipped. In particular, we don't want to make new baselines for android
for non-smoke-tests now since non-smoke-tests are skipped on android.

There's already logic in the copy-existing-baselines to do this; it uses
the TestExpectations.get_expectations method, which, as far as I understand,
returns expectations based on test expectations files, but not based on
whether the port only runs smoke tests.

Meanwhile, elsewhere in rebaseline.py, there's already logic to check
whether a test is skipped on a particular port including because it's not
a smoke test.

The purpose of this CL is to extract that method out so it could
be re-used. An alternative to doing this would be to change
TestExpectations.get_expectations to also include "SKIP" if
port.default_smoke_test_only() and the test is in the smoke tests file.

(Either way, there should be a centralized way of checking whether
a test runs on a port, and this should be used in both places in rebaseline.py.)

BUG=655196

Review-Url: https://codereview.chromium.org/2450383002
Cr-Commit-Position: refs/heads/master@{#428067}