Moves TransientWindowClient to ui/aura/client and adds observer
Mus is going to need to use this from within aura, so I'm moving it to
ui/aura/client. Additionally I'm moving creation of WmState to
AuraTestHelper. Again, mus is going to need it aura test code already
depends upon ui/wm, so I can't think of a reason not to do it
everywhere.

BUG=659155
TEST=covered by tests
R=ben@chromium.org

Review-Url: https://codereview.chromium.org/2453953003
Cr-Commit-Position: refs/heads/master@{#428087}