Remove async load concept for ToS from native code.
This CL removes the async-load concept for ToS from native
ArcSupportHost / ArcAuthService, because it is unused.
Now, it is encapsulated in the arc_support App.

To integrate ToS related code into one place, for the maintainability, this CL introduces
a class, TermsOfServicePage, which implements the ToS page
of the arc_support App.

After this CL, native can just say "Show ToS page" to ArcSupportHost,
so that arc_support app handles async loading.

This is the preparation to resolve the dependency between
ArcAuthService and ArcSupportHost.

BUG=657687
BUG=b/31079732
BUG=b/32424636
TEST=Ran on test device manually. Ran bots.
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2450823002
Cr-Commit-Position: refs/heads/master@{#428375}