Don't run media play promises synchronously if scripts are forbidden.
There are rare situations where we might call the load algorithm while
scripts are not allowed to run. In these cases, we can simply let the
promises resolve later. These are situations where we already do not
conform to the HTML specification and the work around will disappear
when the underlying problem will be fixed.

BUG=657453
R=avayvod@chromium.org, foolip@chromium.org

Review-Url: https://codereview.chromium.org/2459893002
Cr-Commit-Position: refs/heads/master@{#428424}