Stack sampling profiler: temporarily restructure profile read to preserve values for debugging
Minidumps indicate there's something going wrong in the CallStackProfile
Read validation in some cases with injected third party modules, but not
enough information is available to debug the problem. For now, write the
values immediately to the ultimate output, which appears to be preserved
in some dumps. This should be safe since the function still returns
false for invalid data.

BUG=659903

Review-Url: https://codereview.chromium.org/2462783002
Cr-Commit-Position: refs/heads/master@{#428807}