Use DirectX Screen Capturer
This CL makes use of DirectX capturer for screen share.

By testing, we found it is a big performance impact to remove
the "stop sharing" overlay from the captured frame.

After discussing, we decide to just stream everything on the
screen to the remote side. Namely, apply DirectX Capturer
without magnifier.

BUG=314516

Review-Url: https://codereview.chromium.org/2449993004
Cr-Commit-Position: refs/heads/master@{#429086}