Implement CheckExtensionIDs in V4LocalDatabaseManager.
This method checks whether any of the given full hashes is blacklisted in the Chrome Extension blacklist.
It is not being called since the server doesn't support this list for PVer4 yet.
Once the server supports it, I'll add that list to the GetListInfos [1] method for it to start syncing.

[1]: https://cs.chromium.org/chromium/src/components/safe_browsing_db/v4_local_database_manager.cc?q=package:chromium&l=30

This CL is somewhat similar to CL: http://crrev.com/2450413002 which implemented the CheckResourceUrl method.

BUG=543161

Review-Url: https://codereview.chromium.org/2458833004
Cr-Commit-Position: refs/heads/master@{#429090}