Percent height content should respect the height of its containing cell
Another go at the collection of CLs reverted in https://codereview.chromium.org/2309893002.

https://drafts.csswg.org/css-tables-3/#row-layout now says:

"For the purpose of calculating this height, descendants of table cells whose
 height depends on percentages of their parent cell' height are considered to
have an auto height if they have overflow set to *visible or hidden* or if
they are replaced elements, and a 0px height if they have not."

So introduce that behaviour. We also need to account for intrinsic padding and
collapsed border values stored on cells when collecting their border and padding
values.

BUG=637811,627305,623026,465096,468699

Review-Url: https://codereview.chromium.org/2441373002
Cr-Commit-Position: refs/heads/master@{#429109}