Fixed ugly text in the material design system menu.
Labels have subpixel rendering enabled by default which is ugly when
painted on a non-opaque canvas. Currently any system menu label in
a row with a visible ink drop will be given a clear canvas at paint
time because the ink drop hosting layer is promoted to paint to its
layer. This is a short term fix.

BUG=661714

Review-Url: https://codereview.chromium.org/2468173003
Cr-Commit-Position: refs/heads/master@{#429422}