chromeos::NetworkHandler: Add ui_proxy_config_service
This introduces NetworkHandler::InitializePrefServices which
passes PrefService for local (device) state and the logged in user. This will get called twice, once on startup
and once when the primary user logs in.

NetworkHandler::InitializePrefServices currently only initializes
an owned instance of UIProxyConfigService, which can then be used
by any code utilizing NetworkHandler. Currently this is the (old)
options UI, but it will soon include the networkingPrivate
extension API code in order to support the new MD Settings UI.

BUG=658015

Review-Url: https://codereview.chromium.org/2446893008
Cr-Commit-Position: refs/heads/master@{#429528}