[MD History] Fix deletion in heavily scrolled list.
This CL fixes an issue where results would reload when an item was
deleted in a history list that was scrolled beyond the first 'page'.
This was happening because code that refreshed the results on external
deletion was getting triggered for internal deletes due to the web
history service.

This has been fixed by disabling the reload when web history
deletions are occurring, bringing it in line with the local history
service.

BUG=660267, 630164
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2455503004
Cr-Commit-Position: refs/heads/master@{#429529}