Clean up checking Upgrade-Insecure-Requests header
This patch removes checks that the insecure requests header is
AtomicString("1") in favor of just doing a string comparison with "1".

It also removes a duplicate check that is semantically redundant.

Note: for hot code paths the idiom to compare an AtomicString with
AtomicString("<some constant">) is around 6 times slower than just
doing the constant comparison. Code that allocates an AtomicString
constant over and over should consider using statics/constants.

BUG=348655

Review-Url: https://codereview.chromium.org/2454953002
Cr-Commit-Position: refs/heads/master@{#429643}