MacViews: Clear mouse handler when showing context menus.
Currently the View that invokes a menu may incorrectly retain capture and
swallow subsequent mouse events. To fix we explicitly clear mouse handler when
showing context menus.

This same issue was previously fixed individually within ToolbarButton (r15)
and Combobox (r83617) before the common fix inside MenuController (r192688). On
MacViews the MenuController is not used, and we have the MenuRunnerImplCocoa
which didn't contain the fix for this issue.

Move the fix to the common place that's shared between all the implementations
of MenuRunnerImplInterface.

BUG=659204

TEST=Open Chromium, open two tabs with non-empty URLs; Close the second tab;
Right-click the first tab and select Reopen Closed Tab; Click on the second tab.
Second tab should activate. See crbug.com/659204 for video.

Review-Url: https://codereview.chromium.org/2450903002
Cr-Commit-Position: refs/heads/master@{#429673}