Fix pipeline integration tests for seek / duration changes.
PipelineIntegrationTestBase now exposes GetMediaDuration(), so that
tests can seek to the start of the media, rather than assuming it to
be zero.  The ffmpeg_regression_tests would fail, previously, if the
media's start time wasn't zero, since they would seek to 0 and then
get mad that the media time was something else.

Also, this permits the pipeline to notify the client about an
indefinite duration, then notify it about a smaller one.  This
happens if the media duration isn't known initially, but then
becomes known when the end of media is found.  Previously, the
second call to Pipeline::Client::OnDurationChanged would cause
the integration test base to get mad.

Added a test case to ffmpeg_regression_tests which hits both of
these cases.

BUG=662118
TEST=ffmpeg_regression_tests

Review-Url: https://codereview.chromium.org/2479633002
Cr-Commit-Position: refs/heads/master@{#429690}