Add retries for completed background loading attempts
We determined through manual testing that retries are needed for an
acceptable page display rate.  This adds back the retry limit, set to 3
for now, and has the unit test changes to adapt to the retry limit.

Since we now do retries, we set the policy for retries to prefer requests with fewer retries, and among requests with the same number
of retries, to prefer earlier requests.  This will try the requests
in round-robin order so that one troublesome website will not stop
the other requests.

We also removed the check temporarily allowing completed retries (under the guise of started retries)

BUG=660991

Review-Url: https://codereview.chromium.org/2472593002
Cr-Commit-Position: refs/heads/master@{#429748}