Android accessibility: automatically focus links
When setting accessibility focus to a node, auto-focus it if it's a link.
This is needed for some sites that have skip links that are only visible
when focused. It might be possible to auto-focus more than just links but
this seems the safest.

Fixing this exposed a race condition with our code to refocus the node with
accessibility focus if its location has changed - essentially when you
click on the skip link, the skip link was changing its location, causing us
to fire notifications on it because its location had changed - but that
was racing with the real change to accessibility focus from following the
skip link.

Fix this race condition by re-focusing the node with accessibility focus
after a short delay, and canceling the delayed task if accessibility
focus changes again.

Tested manually with two urls that exercise these code paths.

BUG=657157
TESTED=http://getbootstrap.com/ and http://jsfiddle.net/mev0c4dt/show/

Review-Url: https://codereview.chromium.org/2465403003
Cr-Commit-Position: refs/heads/master@{#430006}