Decouple PowerUsageMonitor{Impl} from //content/browser
PowerUsageMonitor is currently only used by Chrome and has significant //content
dependencies, to decouple it we just move it to //chrome/browser at the current
time.
The power_usage_monitor.h is deleted as there is no need to export the
StartPowerUsageMonitor interface from chrome any more. We add the
PowerUsageMonitor as a member into the ChromeBrowserMain.
Unittest are also changed accordingly.

BUG=647247

Review-Url: https://codereview.chromium.org/2456783002
Cr-Commit-Position: refs/heads/master@{#430233}