Remove URLRequest::OrphanJob().
This is a relic of a bygone era, when URLRequestJob was refcounted, and
could outlive its URLRequest. Now we can just delete the URLRequestJob.

Also remove a couple checks in URLRequestHttpJob and
ServiceWorkerURLRequestJob that their request is not NULL, as the checks
are no longer needed, for the same reason.

BUG=NONE

Review-Url: https://codereview.chromium.org/2480563002
Cr-Commit-Position: refs/heads/master@{#430301}