[ObjC ARC] Converts crw_web_controller to ARC.
Notable changes:
* crw_web_controller.mm used to have a dictionary storing selectors just
for keeping the KVO code clean. ARC cannot call a selector from a string
like this, because it cannot guarantee correct memory management in
this case. So it was replaced with an array of observed key paths to
keep some of the code cleanness, while the KVO callback has a huge
switch.

BUG=624365
TEST=None

Review-Url: https://codereview.chromium.org/2434853002
Cr-Commit-Position: refs/heads/master@{#430304}