mus: Remove unnecessary GpuMemoryBuffer code from client-lib.
The gfx::GpuMemoryBuffer implementations in the mus client lib
do not do the right thing, and are not really necessary. Remove
the code, and instead use the same implementation in gpu-client
that content uses.

The code is still not right: instead of getting memory buffer
from the host, the code allocates shared memory on the client
side. When returning memory, the code ignores the sync token.
These issues will be fixed in future CLs.

BUG=643746

Review-Url: https://codereview.chromium.org/2480023002
Cr-Commit-Position: refs/heads/master@{#430335}