Don't initialize ICU replacement character for UTF-32 encodings
The TextCodecICU wrapper erroneously called ucnv_setSubstChars() with
only a single byte replacement for non-byte-based encodings -
basically the UTF-32 family. The code path is only exercised when
computing names for files in form uploads. This would hit an ASSERT in
debug builds, and result in an empty filename in release builds.

BUG=662146,417850
R=tkent@chromium.org

Review-Url: https://codereview.chromium.org/2478653002
Cr-Commit-Position: refs/heads/master@{#430356}