[TTS] Change layout of short text in Bar for RTL.
Adds a new class OverlayPanelTextViewInflater that automatically adjusts
a TextView after inflation to override RTL/LTR ordering when the text is
considered "short".  Update the TTS Bar Text and Caption to use this.

When we allow the natural RTL and LTR layout in the Bar, with translation
enabled we often see the caption going one way and the main bar text going
the other way, which looks broken.  We solve this by checking
if the text is short and forcing a match to the layout of the UI
language. This is done for both pieces of text independently.

BUG=651389

Review-Url: https://codereview.chromium.org/2477053002
Cr-Commit-Position: refs/heads/master@{#430420}