Views:: Make Labels support text selection.
This CL is a follow-up to r427604 which extracted the text selection logic from
views::Textfield and introduced the SelectionController class. This CL makes
views::Label implement the SelectionControllerDelegate interface. As a result,
single-line views::Label now support text selection.

Some of the member functions in the Label class which do not change the external
state of the Label are made const. This is made possible by making the render
text instances that are used for drawing, mutable.

A new test fixture class is also added for testing Labels. Existing tests are
modified to use it, and new tests are added to test the text selection behavior.

Support for copying and dragging selected text and text selection in multi-line
views::Label will be added subsequently.

Doc= https://docs.google.com/a/google.com/document/d/1JTjY8RLE84hxubCAouHuD6xVtvuMnzbqSDZyf_4N6tY/edit?usp=sharing
BUG=630365, 437993
TEST= Launch views_examples_with_content_exe. Select Labels from the dropdown.
Verify that text selection in the label with the text "A label supporting text
selection." works properly. On Mac, also pass the --enable-harfbuzz-rendertext
flag.

Review-Url: https://codereview.chromium.org/2413223003
Cr-Commit-Position: refs/heads/master@{#430461}