[DiskCache] Create SimpleCache Size Experiment
This CL does the following:

1) Creates a bit of infrastructure for running SimpleCache experiments.

A cache has a single experiment which is written in its FakeIndex. The default
experiment is NONE. If, when opening the FakeIndex, it's discovered that the
FakeIndex is of the wrong experiment type, the cache will be destroyed and
remade with the correct type. When the experiment is over the user will rejoin
the NONE experiment and reset their cache once again.

2) Creates a simple SIZE experiment for altering the size of the cache

After determining what the cache size should be, multiplies it by a factor
gleaned from a field trial parameter.

BUG=

Review-Url: https://codereview.chromium.org/2480553002
Cr-Commit-Position: refs/heads/master@{#430625}